Update FindPyCXX.cmake for Fedora and 7.x
This specific commit probably requires some more modification/thought
to not break other use cases.

However, for now, this brings Fedora builds up to date to use external
PyCXX 7.x without compatability mode, as well as actually *find* the
PyCXX install on contemporary Fedora.

Specifically, look for it in the right place and do not
force-hardwaire to bundled (out of date PyCXX 6.1)